主题添加市场开放时间及可转售/注册拉新

+++ star-domain/src/main/java/com/starnft/star/domain/rank/core/rank/model/res/RankingsItem.java
@@ -18,7 +18,7 @@ public class RankingsItem {
     @ApiModelProperty(value = "账号")
     private Long account;
     @ApiModelProperty(value = "总人数")
-    private Integer total;
+    private Long total;
     @ApiModelProperty(value = "有效人数")
-    private Integer valid;
+    private Long valid;
 }
